Putting images back.
When using docs-resources, adding images to the docs-resources/images directory doesn't make sense, as then you'd have to
add the to the submodule upstream to make them available to others.  Thererfore, keep images in root images directory, set it as the imagesdir and just reference the docs-resources/images dir for the logo.